Graphic : Update

+++ xxGraphic.h
@@ -62,7 +62,7 @@ xxAPI void          (*xxUnmapBuffer)(uint64_t device, uint64_t buffer);
 //==============================================================================
 //  Texture
 //==============================================================================
-xxAPI uint64_t      (*xxCreateTexture)(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external);
+xxAPI uint64_t      (*xxCreateTexture)(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external);
 xxAPI void          (*xxDestroyTexture)(uint64_t texture);
 xxAPI void*         (*xxMapTexture)(uint64_t device, uint64_t texture, int* stride, int level, int array);
 xxAPI void          (*xxUnmapTexture)(uint64_t device, uint64_t texture, int level, int array);
