+++ xxGraphicD3D11.cpp
@@ -710,7 +710,7 @@ void xxUnmapBufferD3D11(uint64_t device, uint64_t buffer)
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t xxCreateTextureD3D11(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external)
+uint64_t xxCreateTextureD3D11(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external)
 {
     ID3D11Device* d3dDevice = reinterpret_cast<ID3D11Device*>(device);
     if (d3dDevice == nullptr)
