+++ xxGraphicD3D9.cpp
@@ -531,7 +531,7 @@ void xxUnmapBufferD3D9(uint64_t device, uint64_t buffer)
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t xxCreateTextureD3D9(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external)
+uint64_t xxCreateTextureD3D9(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external)
 {
     LPDIRECT3DDEVICE9 d3dDevice = reinterpret_cast<LPDIRECT3DDEVICE9>(device);
     if (d3dDevice == nullptr)
