+++ xxGraphicGlide.h
@@ -66,7 +66,7 @@ void            xxUnmapBufferGlide(uint64_t device, uint64_t buffer);
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t        xxCreateTextureGlide(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external);
+uint64_t        xxCreateTextureGlide(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external);
 void            xxDestroyTextureGlide(uint64_t texture);
 void*           xxMapTextureGlide(uint64_t device, uint64_t texture, int* stride, int level, int array);
 void            xxUnmapTextureGlide(uint64_t device, uint64_t texture, int level, int array);
