+++ xxGraphicMantle.h
@@ -66,7 +66,7 @@ void            xxUnmapBufferMantle(uint64_t device, uint64_t buffer);
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t        xxCreateTextureMantle(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external);
+uint64_t        xxCreateTextureMantle(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external);
 void            xxDestroyTextureMantle(uint64_t texture);
 void*           xxMapTextureMantle(uint64_t device, uint64_t texture, int* stride, int level, int array);
 void            xxUnmapTextureMantle(uint64_t device, uint64_t texture, int level, int array);
