+++ xxGraphicVulkan.h
@@ -66,7 +66,7 @@ void            xxUnmapBufferVulkan(uint64_t device, uint64_t buffer);
 //==============================================================================
 //  Texture
 //==============================================================================
-uint64_t        xxCreateTextureVulkan(uint64_t device, int format, int width, int height, int depth, int mipmap, int array, void const* external);
+uint64_t        xxCreateTextureVulkan(uint64_t device, uint64_t format, int width, int height, int depth, int mipmap, int array, void const* external);
 void            xxDestroyTextureVulkan(uint64_t texture);
 void*           xxMapTextureVulkan(uint64_t device, uint64_t texture, int* stride, int level, int array);
 void            xxUnmapTextureVulkan(uint64_t device, uint64_t texture, int level, int array);
